requirements.txt
SQLAlchemy~=2.0.36
fastapi~=0.115.4
alembic~=1.13.3
python-dotenv~=1.0.1
jinja2~=3.1.4